event-loop-base: fix property typenames
The getter and setter use visit_type_int64, not visit_type_int.

Fixes: 7d5983e3c8c4 ("Introduce event-loop-base abstract class")
Message-Id: <20260818-qom-qapi-v3-35-24b8bbbe3d86@redhat.com>

+++ event-loop-base.c
@@ -104,15 +104,15 @@ static void event_loop_base_class_init(ObjectClass *klass,
     ucc->complete = event_loop_base_complete;
     ucc->prepare_delete = event_loop_base_prepare_delete;
 
-    object_class_property_add(klass, "aio-max-batch", "int",
+    object_class_property_add(klass, "aio-max-batch", "int64",
                               event_loop_base_get_param,
                               event_loop_base_set_param,
                               NULL, &aio_max_batch_info);
-    object_class_property_add(klass, "thread-pool-min", "int",
+    object_class_property_add(klass, "thread-pool-min", "int64",
                               event_loop_base_get_param,
                               event_loop_base_set_param,
                               NULL, &thread_pool_min_info);
-    object_class_property_add(klass, "thread-pool-max", "int",
+    object_class_property_add(klass, "thread-pool-max", "int64",
                               event_loop_base_get_param,
                               event_loop_base_set_param,
                               NULL, &thread_pool_max_info);
